requirements.txt
# ╔════════════════════════════════════════════════════════════════════════════╗
# ║ NoteVibe – Python Requirements File                                      ║
# ║ Purpose: List of all Python dependencies for the project.                ║
# ╚════════════════════════════════════════════════════════════════════════════╝
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.0.1